Move cuda_visible_devices to utils.py
